ticket_gym/requirements.txt
Django==4.2.8
djangorestframework==3.15.2
psycopg2-binary==2.9.8
channels==4.2.0.
channels_redis==4.0.0.
asyncpg==0.27.0.
docker-compose==2.29.2

